app_2/llm/requirements.txt
# LLM Scheduler Requirements
openai>=1.0.0
python-dotenv>=1.0.0
backoff>=2.2.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0